feat(executor): 62V6Q5 — facts-driven Plan walker, tri-state derivation, pack_config dedup
Close out epic 62V6Q5 ("Executor deepening"): the 30 hand-written
build_*_plan dispatch table is dissolved into a single facts-driven
Plan walker (ADR-031); the derivation surface distinguishes a routine
Decline from a validator Reject that is always fatal (ADR-030); and
pack_config is deduplicated onto config.rs as the single canonical home.

Tasks:
- B (6CXN2R): enum Derive { Encoded | Declined | Rejected(ValidationError) }
  at the derivation layer; encode_grammar/encode_cmd_stream stay
  Option<Vec<u8>> (Declined -> None); a built-then-rejected Plan panics
  with the invariant + offending op, never a silent skip.
- A1 (3LLE7X): spike proving the ADR-031 hybrid schemas (per-protocol
  hop facts + mechanics + enclosure-deriving walker) express v3_v4_v3
  execution-identical to build_v3v4v3_plan through the revm matrix.
- A2 (AR5TJN): walker generalized to every family behind the walk
  feature; family_axis_support is facts-derived, not a 30-row table.
- A3 (KFFR6Q): walk is the default/sole producer; the 30 build_*_plan
  bodies and the build_for dispatch shell are deleted; grammar_shape.rs
  shrinks to the thin shell (~1624 lines) + grammar_walker.rs (the
  ~5070-line facts-driven walker) + grammar_plan.rs/grammar_ledger.rs.
  build_for is gone; both call sites (derive_shape_detailed,
  family_axis_support) inline grammar_walker::build_for_walk directly.
- C (KKQTWZ): config.rs owns pack_config/pack_expected_balance with a
  single ConfigError, bribe_bips: u16, MAX_BRIBE_BIPS=10_000;
  encoders.rs re-exports rather than re-implementing, and the three
  EncoderError::{InvalidCheckMode,InvalidBribeBips,InvalidBribeRecipientIdx}
  variants are removed.

Gates (full A3 combined suite, all green):
  just test-rust, just lint-rust, just check-no-pyo3-in-cores,
  just test-python (2498 passed, 30 skipped).

ADRs: ADR-030 (derivation outcome tri-state), ADR-031 (executor Plan
walker). CONTEXT.md gains the "Derivation outcome", "Hop facts", and
"Enclosure (derived)" terms.

Note: A1's facts-schema checkpoint was generalized to A2 without a
recorded user sign-off; the schema is proven by the full revm matrix
+ golden corpora parity (A2's gate), so this is recorded as a
process note rather than a blocker.

+++ rust/crates/degenbot-executor/src/config.rs
@@ -17,7 +17,7 @@ use alloy::primitives::U256;
 use crate::encoders::MAX_INDEXED_ADDRESSES;
 
 /// Max bribe in basis points (100% = 10000 bips).
-pub const MAX_BRIBE_BIPS: u32 = 10_000;
+pub const MAX_BRIBE_BIPS: u16 = 10_000;
 
 /// Max `check_mode` selector (0=skip, 1=WETH+ETH, 2=ERC6909 WETH, 3=SWEEP).
 pub const MAX_CHECK_MODE: u8 = 3;
@@ -30,7 +30,7 @@ pub enum ConfigError {
     CheckModeOutOfRange(u8),
     /// `bribe_bips` not in `0..=10000`.
     #[error("bribe_bips must be 0–10000, got {0}")]
-    BribeBipsOutOfRange(u32),
+    BribeBipsOutOfRange(u16),
     /// `bribe_recipient_idx` not in `0..32`.
     #[error("bribe_recipient_idx must be 0–31, got {0}")]
     BribeRecipientOutOfRange(u8),
@@ -50,7 +50,7 @@ pub enum ConfigError {
 pub fn pack_config(
     check_mode: u8,
     expected_value: U256,
-    bribe_bips: u32,
+    bribe_bips: u16,
     bribe_recipient_idx: u8,
 ) -> Result<U256, ConfigError> {
     if check_mode > MAX_CHECK_MODE {
